WIP use hook instead of RPC for sync
This basically works, but it still has the problem that once we
complete a job, we don't really want to do another until we've seen
the commit we just pushed in the main repo.

Before, we could instruct the repo to pull from upstream
synchronously, then proceed. Now, it's unclear how to do this because
the repo might not be ready.